requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
psycopg2-binary>=2.9,<3.0
redis>=5.0,<6.0
python-decouple>=3.8,<4.0
